+++ navigation/atlassian-switcher/src/create-custom-provider.ts
@@ -2,3 +2,4 @@ export { createAvailableProductsProvider } from './common/providers/default-avai
 export { createJoinableSitesProvider } from './cross-join/providers/default-joinable-sites-provider';
 export { defaultFetchData as defaultJoinableSitesFetch } from './cross-join/providers/default-joinable-sites-fetch';
 export { fetchProductRecommendations } from './cross-join/providers/product-recommendations-fetch';
+export { productRecommendationsNonAAFetchData } from './cross-join/providers/product-recommendations-non-aa-fetch';