+++ people-and-teams/profilecard/src/api/index.ts
@@ -1,0 +1,8 @@
+import CachingClient from './CachingClient';
+import ProfileCardClient from './ProfileCardClient';
+import TeamProfileClient from './TeamProfileCardClient';
+import UserProfileClient, { modifyResponse } from './UserProfileCardClient';
+
+export { CachingClient, modifyResponse, TeamProfileClient, UserProfileClient };
+
+export default ProfileCardClient;