+++ people-and-teams/profilecard/src/components/user.ts
@@ -1,0 +1,7 @@
+import { DELAY_MS_HIDE, DELAY_MS_SHOW } from './config';
+import ProfileCard from './ProfileCard';
+import ProfileCardTrigger from './ProfileCardTrigger';
+
+export { DELAY_MS_HIDE, DELAY_MS_SHOW, ProfileCard, ProfileCardTrigger };
+
+export default ProfileCardTrigger;