+++ editor/editor-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-common
 
+## 55.2.1
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 55.2.0
 
 ### Minor Changes